+++ snuba/datasets/cdc/groupedmessage.py
@@ -5,7 +5,7 @@ from snuba.clickhouse.columns import ColumnSet, DateTime, Nullable, UInt
 from snuba.datasets.cdc import CdcDataset
 from snuba.datasets.dataset_schemas import DatasetSchemas
 from snuba.datasets.cdc.groupedmessage_processor import GroupedMessageProcessor, GroupedMessageRow
-from snuba.datasets.schema import ReplacingMergeTreeSchema
+from snuba.datasets.schemas.table_schemas import ReplacingMergeTreeSchema
 from snuba.snapshots.bulk_load import SingleTableBulkLoader
