+++ snuba/datasets/schemas/join_schema.py
@@ -3,7 +3,7 @@ from enum import Enum
 from typing import Callable, Mapping, Optional, Sequence
 
 from snuba.clickhouse.columns import ColumnSet
-from snuba.datasets.schema import Schema
+from snuba.datasets.schemas import Schema
 
 
 class JoinType(Enum):
@@ -71,12 +71,10 @@ class JoinedSchema(Schema):
 
     def __init__(self,
         join_root: JoinStructure,
-        migration_function: Optional[Callable[[str, Mapping[str, str]], Sequence[str]]]=None,
     ) -> None:
         self.__join_storage = join_root
         super().__init__(
             columns=ColumnSet([]),  # TODO: process the joined table to build the columns list
-            migration_function=migration_function
         )
 
     def get_columns(self):
