+++ snuba/datasets/transactions.py
@@ -17,7 +17,7 @@ from snuba.clickhouse.columns import (
 )
 from snuba.datasets import TimeSeriesDataset
 from snuba.datasets.dataset_schemas import DatasetSchemas
-from snuba.datasets.schema import ReplacingMergeTreeSchema
+from snuba.datasets.schemas.table_schemas import ReplacingMergeTreeSchema
 from snuba.datasets.transactions_processor import TransactionsMessageProcessor
 from snuba.query.extensions import PERFORMANCE_EXTENSION_SCHEMA, PROJECT_EXTENSION_SCHEMA
 from snuba.query.schema import GENERIC_QUERY_SCHEMA
